Handle Cloudflare protection challenge.
Cloudflare protection mitigates automated attacks by intercepting
requested content, issuing an HTTP response with status code 403, and
and page content that requires either human input or browser evaluation
of Javascript challenge code.

This update adds logic to recognize the Javascript challenge, by
rendering the response content and thereby cause the browser to execute
the Cloudflare challenge Javascript code.  This permits an operator to
react to a challenge directed at a human operator, such as a page
requiring them to click on a checkbox in order to proceed.

The challenge handler logic does not yet resume an interrupted page
fetch operation.